so, apparently there is still a bug for a very special case, where the first connection is at the vertex at the end of the list... I think I can fix it tomorrow, but right now I'm too tired. But at least it works for the case of two obstacles :) So when you just run this without changing anything, you can see the two splitted polygons and the triangles including midpoints To use the midpoints you have the variable midpoints in the class Floor
I don't have the updated version of esthers adjaceny method for the midpoints, so you would need to copy that manually...
As an input it should have the midpoints list, which includes the midpoints of both of the splitted polygons.

With this we have at least something to record and show :)

Oh and it can also be tried out if there is a difference if we use the ratio test for the triangulation or not by changing the name of the function from getTriangulation(...) to getTriangulationWithRatioTest(...) in lines 865 and 922.
I don't know if it makes a huge difference for this example but I saw that it differs for the map with 5 obstacles when I just used one splitted polygon.